requirements.txt
# --- Job Hunter v2 Dependencies ---

# Core Framework
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-anthropic>=0.1.20

# Messaging
python-telegram-bot>=21.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
httpx>=0.27.0
rich>=13.7.0
aiosqlite>=0.20.0
deepagents>=0.4.0
aiosmtplib>=3.0.1
google-genai>=1.24.0

# Optional: Data Analysis (if needed by skills)
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
langchain-google-genai
